+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0]
+
+### Added
+- Added the capacity to replicate assets' folder structure into collections.
+
+### Changed
+- Updated dependency installation process to use requirements.txt
+
 ## [0.4.0] - 2025-02-28
 
 ### Added